Make filename and path strings longer to avoid truncation
This commit creates a str_len parameter of 200 characters and uses it to
define the length of all string variables used for filename or path
names.  This avoids filename truncation and associated read errors when
long filenames are used.  It also fixes the fact that the 'times' output
file had been shortened to 'time'.  Those truncation issues were
introduced in the switch to using a namelist, such that the associated
string variables could not be defined with unspecified length (*).