Change `norder` and `nglv` to namelist options
Previously, the variables `norder` and `nglv` were parameters that
were used in declaring the dimension of many variables. This has raised
two issues in our recent efforts to run a suite of sensitivity coupled MALI-SLM
simulations that sweeps across various values of MALI mesh resolution
and the number of Gauss-Legendre nodes in latitude (represented by variable `nglv`)
in the SLM. The first issue (or more like an inconvenience) was the need to code compilation needs to be done
everytime `nglv` parameter was changed, and the second was the failure
of compilation when the `nglv` value was set higher than 2000 because of limited memory imposed to
variables that are declared with defined a dimension size.
This PR addresses the two issues by 1) no longer assigning the variables `norder` and `nglv` as
model parameters, and 2) making the variables that use these variables to declare their dimension as allocatable.